dean/src/infrastructure/performance_optimization.py
"""
Performance Optimization
Database query optimization, image compression, parallel processing
"""

from __future__ import annotations

import logging
import asyncio
from typing import Any, Callable, Dict, List, Optional
from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
import functools

logger = logging.getLogger(__name__)


class AsyncProcessor:
    """Async processing for non-blocking operations."""
    
    def __init__(self, max_workers: int = 5):
        self.executor = ThreadPoolExecutor(max_workers=max_workers)
        self.loop = None
    
    def process_async(
        self,
        func: Callable,
        *args,
        **kwargs,
    ):
        """Process function asynchronously."""
        future = self.executor.submit(func, *args, **kwargs)
        return future
    
    def process_batch(
        self,
        items: List[Any],
        func: Callable,
        max_workers: Optional[int] = None,
    ) -> List[Any]:
        """Process batch of items in parallel."""
        with ThreadPoolExecutor(max_workers=max_workers or 5) as executor:
            futures = [executor.submit(func, item) for item in items]
            return [f.result() for f in futures]


class ImageOptimizer:
    """Image optimization and compression."""
    
    def __init__(self):
        self.quality = 85  # JPEG quality
        self.max_size_mb = 5.0
    
    def optimize_image(
        self,
        image_path: str,
        output_path: Optional[str] = None,
    ) -> str:
        """Optimize image."""
        try:
            from PIL import Image
            
            img = Image.open(image_path)
            
            # Convert to RGB if needed
            if img.mode != 'RGB':
                img = img.convert('RGB')
            
            # Resize if too large
            max_dimension = 2048
            if max(img.size) > max_dimension:
                ratio = max_dimension / max(img.size)
                new_size = tuple(int(d * ratio) for d in img.size)
                img = img.resize(new_size, Image.Resampling.LANCZOS)
            
            # Save optimized
            output = output_path or image_path
            img.save(
                output,
                'JPEG',
                quality=self.quality,
                optimize=True,
            )
            
            logger.info(f"Optimized image: {image_path}")
            return output
        
        except Exception as e:
            logger.error(f"Image optimization error: {e}")
            return image_path


class DatabaseQueryOptimizer:
    """Database query optimization."""
    
    def __init__(self):
        self.query_cache: Dict[str, Any] = {}
        self.slow_queries: List[Dict[str, Any]] = []
    
    def optimize_query(
        self,
        query: str,
        params: Dict[str, Any] = None,
    ) -> str:
        """Optimize database query."""
        # Add query hints, indexes, etc.
        # Simplified implementation
        optimized = query
        
        # Add LIMIT if not present and query could be large
        if "LIMIT" not in query.upper() and "SELECT" in query.upper():
            optimized = f"{query} LIMIT 1000"
        
        return optimized
    
    def batch_insert(
        self,
        table: str,
        records: List[Dict[str, Any]],
        batch_size: int = 100,
    ) -> List[Any]:
        """Batch insert records."""
        results = []
        
        for i in range(0, len(records), batch_size):
            batch = records[i:i + batch_size]
            # Insert batch
            # Implementation depends on database
            results.append(batch)
        
        return results


def parallel_process(
    items: List[Any],
    func: Callable,
    max_workers: int = 5,
) -> List[Any]:
    """Process items in parallel."""
    processor = AsyncProcessor(max_workers=max_workers)
    return processor.process_batch(items, func, max_workers)


def optimize_image(image_path: str) -> str:
    """Optimize image."""
    optimizer = ImageOptimizer()
    return optimizer.optimize_image(image_path)


__all__ = [
    "AsyncProcessor",
    "ImageOptimizer",
    "DatabaseQueryOptimizer",
    "parallel_process",
    "optimize_image",
]

