dean/requirements.txt
# Core Dependencies
facebook-business==19.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
pytz>=2020.1,<2024
pandas==2.2.2
openpyxl==3.1.5
SQLAlchemy==2.0.35
prometheus-client==0.20.0
jsonschema==4.23.0
supabase>=2.5.0
schedule==1.2.0

# Core ML Dependencies (Essential)
xgboost>=1.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
joblib>=1.3.0

# Important ML Dependencies (Recommended)
lightgbm>=4.0.0
catboost>=1.2.0
statsmodels>=0.14.0
psycopg2-binary>=2.9.0

# Visualization (Recommended for ML insights)
plotly>=5.17.0
seaborn>=0.13.0
matplotlib>=3.8.0

# Time Series ML (Important for temporal modeling)
tslearn>=0.6.0
prophet>=1.1.0
arch>=6.0.0

# Advanced ML Features (Optional but useful)
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
dask>=2023.0.0
optuna>=3.4.0
redis>=5.0.0  # Caching infrastructure
psutil>=5.9.0  # For resource optimization
alembic>=1.13.0

# ML Experiment Tracking (Optional)
mlflow>=2.8.0
wandb>=0.16.0
tensorboard>=2.15.0

# Advanced ML Features (NEW - for 20 improvements)
torch>=2.0.0  # Neural networks (LSTM)
shap>=0.42.0  # Model explainability
featuretools>=1.26.0  # Automated feature engineering
gplearn>=0.4.2  # Genetic programming features
scikit-optimize>=0.9.0  # Bayesian optimization
imbalanced-learn>=0.11.0  # Handling imbalanced data
pyod>=1.1.0  # Outlier/anomaly detection

# Creative Intelligence System Dependencies
openai>=1.0.0  # GPT-5 integration
sentence-transformers>=2.2.0  # Semantic similarity analysis
transformers>=4.30.0  # Advanced NLP models
spacy>=3.6.0  # Advanced NLP
nltk>=3.8.0  # Natural language processing
Pillow>=10.0.0  # Image processing for creative analysis and optimization
opencv-python>=4.8.0  # Image processing
textblob>=0.17.0  # Text analysis
vaderSentiment>=3.3.0  # Sentiment analysis

# Image Generation & Processing
ffmpeg-python>=0.2.0  # ffmpeg wrapper for text overlay